Update sequence number in source handler
